refactor(util/pattern-viewer): Move pattern extraction to new file, rename imports
