+++ lib/util/viewer.js
@@ -97,57 +97,6 @@ export function extractHeadsignFromPattern(pattern, routeShortName = null) {
   return headsign
 }
 
-export function extractMainHeadsigns(patterns, shortName, editHeadsign) {
-  const mapped = Object.entries(patterns).map(
-    ([id, pat]) /* : PatternSummary */ => ({
-      geometryLength: pat.patternGeometry?.length || 0,
-      headsign: extractHeadsignFromPattern(pat, shortName),
-      id,
-      lastStop: pat.stops?.[pat.stops?.length - 1]?.name
-    })
-  )
-
-  // Address duplicate headsigns.
-  return mapped.reduce((prev /* : PatternSummary[] */, cur) => {
-    const amended = prev
-    const alreadyExistingIndex = prev.findIndex(
-      (h) => h.headsign === cur.headsign
-    )
-    // If the headsign is a duplicate, and the last stop of the pattern is not the headsign,
-    // amend the headsign with the last stop name in parenthesis.
-    // e.g. "Headsign (Last Stop)"
-    if (
-      alreadyExistingIndex >= 0 &&
-      cur.lastStop &&
-      cur.headsign !== cur.lastStop
-    ) {
-      editHeadsign(cur)
-
-      // If there are only two total patterns, then we should rename
-      // both of them
-      if (amended.length === 1 && Object.entries(patterns).length === 2) {
-        editHeadsign(amended[0])
-        amended.push(cur)
-        return amended
-      }
-    }
-
-    // With all remaining duplicate headsigns with the same last stops, only keep the pattern with the
-    // longest geometry.
-    if (
-      alreadyExistingIndex >= 0 &&
-      amended[alreadyExistingIndex].lastStop === cur.lastStop
-    ) {
-      if (amended[alreadyExistingIndex].geometryLength < cur.geometryLength) {
-        amended[alreadyExistingIndex] = cur
-      }
-    } else {
-      amended.push(cur)
-    }
-    return amended
-  }, [])
-}
-
 /**
  * Gets the mode string from either an OTP Route or RouteShort model. The OTP
  * Route model returns the mode as an integer type whereas the RouteShort model
